Reorganize Python tests
