+++ build.gradle
@@ -204,11 +204,12 @@ test.dependsOn(checkSettings)
 
 task testPython(type: Exec, dependsOn: shadowJar) {
     commandLine 'pytest',
+            '-v',
             '-n',
             parallelism,
             '--dist=loadscope',
             '--noconftest',
-            'python/hail/tests'
+            'python/tests'
     environment SPARK_HOME: sparkHome
     environment PYTHONPATH: '' + projectDir + '/python:' + sparkHome + '/python:' + sparkHome + '/python/lib/py4j-' + py4jVersion + '-src.zip'
     environment PYSPARK_SUBMIT_ARGS: '--conf spark.driver.extraClassPath=' + projectDir + '/build/libs/hail-all-spark.jar --conf spark.executor.extraClassPath=' + projectDir + '/build/libs/hail-all-spark.jar pyspark-shell'
