+++ python/tests/test_hail/genetics/test_locus.py
@@ -1,0 +1,17 @@
+import unittest
+
+import hail as hl
+from hail.genetics import *
+from ..helpers import *
+
+setUpModule = startTestHailContext
+tearDownModule = stopTestHailContext
+
+class Tests(unittest.TestCase):
+
+    def test_constructor(self):
+        l = Locus.parse('1:100')
+
+        self.assertEqual(l, Locus('1', 100))
+        self.assertEqual(l, Locus(1, 100))
+        self.assertEqual(l.reference_genome, hl.default_reference())
